requirements.txt
braintrust[otel]
crewai
opentelemetry-instrumentation-openai
opentelemetry-instrumentation-crewai
python-dotenv
